+++ demo/pg_create.go
@@ -365,9 +365,7 @@ func CheckSelectedCluster() {
 
 	makeup.Print("The currently selected Kubernetes context is: " + current_context)
 
-	var desired_context_name string
-
-	desired_context_name = DemoClusterName
+	desired_context_name := GetKubernetesCreator().GetContext(DemoClusterName)
 
 	if strings.HasPrefix(current_context, desired_context_name) {
 		makeup.PrintCheckmark("It seems that the right context is selected: " + desired_context_name)